Refactor adapters
Separate into different files; Uses es6 class inheritance instead of utility functions;
