+++ web_external/views/adapters/ContourRepresentation.js
@@ -1,0 +1,12 @@
+import GeometryRepresentation from './GeometryRepresentation';
+
+/**
+ * Generic GeoJs Contour MapRepresentation definition, with type 'contour'.
+ */
+class ContourRepresentation extends GeometryRepresentation {
+    constructor() {
+        super();
+        this.readerType = 'contourJsonReader';
+    }
+}
+export default ContourRepresentation;
