+++ web_external/views/map/MapPanel.js
@@ -2,7 +2,7 @@ import _ from 'underscore';
 import geo from 'geojs';
 
 import Panel from '../body/Panel';
-import adapterRegistry from '../adapters/Adapters';
+import adapterRegistry from '../adapters/registry';
 import FeatureInfoWidget from '../widgets/FeatureInfoWidget';
 import template from '../../templates/body/mapPanel.pug';
 import '../../stylesheets/body/mapPanel.styl';
